requirements.txt
selenium
pymongo
flask
requests
python-dotenv